browseragent2/pyproject.toml
[project]
name = "browseragent2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "stagehand>=0.4.0",
    "python-dotenv>=1.0.0",
]
